Obtain OG from Gene

This function retrieves the OG (Orthogroup) and Gene Tree Parent Clade for a given GeneID and SpeciesID. It takes a GeneDiscoveR object, a GeneID, and a SpeciesID as input and returns a data frame containing the OG, Gene Tree Parent Clade, and GeneID for the matching GeneID and SpeciesID. If no matching GeneID and SpeciesID are found, the function returns NA for the OG and Gene Tree Parent Clade.

obtain_OG_from_gene(GeneDiscoveRobject, GeneID = NULL, SpeciesID = NULL)

Arguments

GeneDiscoveRobject

The GeneDiscoveR object containing the complete table.

GeneID

A character vector of GeneIDs.

SpeciesID

A character vector of SpeciesIDs.

Value

A data frame containing the OG, Gene Tree Parent Clade, and GeneID for the matching GeneIDs and SpeciesIDs.

Examples

N0Dir <- system.file("extdata", "Brassicaceae", package = "GeneDiscoveR")
dataTSV <- system.file("extdata", "Brassicaceae", "table_traits_selfcomp.tsv", package = "GeneDiscoveR")

# Create a GeneDiscoveR object with Brassicaceae data from one execution of OrthoFinder
# In this case, because we are unique execution, we use the uniqueInflation parameter and GeneDiscoveR set automatically the active run.
GeneDiscoveRobject <- GeneDiscoveR(
    N0sDir = N0Dir,
    dataFile = dataTSV,
    uniqueInflation = 1.5,
    orthologsTool = "OrthoFinder"
)
#> -----------From OrthoFinder-----------
#> The process has been completed successfully

# Select species by phenotype. You can perform this step with different phenotypes.
GeneDiscoveRobject <- select_species_by_phenotype(
    GeneDiscoveRobject = GeneDiscoveRobject,
    columnPhenotype = "Self-compatible",
    columnID = "OrthofinderID",
    type = "0"
)
GeneDiscoveRobject <- select_species_by_phenotype(
    GeneDiscoveRobject = GeneDiscoveRobject,
    columnPhenotype = "Self-compatible",
    columnID = "OrthofinderID",
    type = "1"
)

# Identify genes by phenotype. You can perform this step with different phenotypes.
GeneDiscoveRobject <- gene_identification_by_phenotype(
    GeneDiscoveRobject = GeneDiscoveRobject,
    formula = as.formula("1 ~ 0"),
    statistic = "Fisher",
    name = "Self-incompatible",
    cores = 8
)
GeneDiscoveRobject <- select_genes_by_phenotype(GeneDiscoveRobject,
    pvalue = 0.05,
    oddsRatio = 1,
    sign = ">",
    name = "Self-incompatible"
)

# Set annotation file
# Import Arabidopsis thaliana annotation file from TAIR10
annotationFile <- system.file("extdata", "Brassicaceae", "TAIR10_functional_descriptions", package = "GeneDiscoveR")

GeneDiscoveRobject <- set_annotation_file(GeneDiscoveRobject, annotationFile = annotationFile)
indexFilteredGenes <- select_filtered_gene_index(GeneDiscoveRobject, name = "Self-incompatible", pvalue = 0.05, oddsRatio = 1, sign = ">")

# Map annotation to the filtered genes. indexFilteredGenes is the index of the filtered genes, if NULL, the annotation is mapped to the complete table
GeneDiscoveRobject <- map_annotation(
    GeneDiscoveRobject = GeneDiscoveRobject,
    indexFilteredGenes = indexFilteredGenes,
    specieWithAnnotation = "Athaliana_447_Araport11.protein_primaryTranscriptOnly",
    oneColumn = FALSE
)
#> Rows: 41672 Columns: 5
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (5): X1, X2, X3, X4, X5
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

# Show the filtered genes table
head(get_filtered_genes_table(GeneDiscoveRobject, name = "Self-incompatible", pvalue = 0.05, oddsRatio = 1, sign = ">"))
#>          OG Gene Tree Parent Clade           HOG
#> 1 OG0000007                  n1197 N0.HOG0000020
#> 2 OG0000183                    n18 N0.HOG0000303
#> 3 OG0000495                     n0 N0.HOG0000725
#> 4 OG0000789                     n1 N0.HOG0001112
#> 5 OG0001631                     n0 N0.HOG0002136
#> 6 OG0002789                     n1 N0.HOG0003391
#>                                                                                                                                                                                      Alinifolium_472_v1.1.protein_primaryTranscriptOnly
#> 1                                                                                                                                                                                                                                  <NA>
#> 2                                                                                                                                                                                                                                  <NA>
#> 3 Alyli.0042s0073.1.p, Alyli.0060s0092.1.p, Alyli.0480s0012.1.p, Alyli.0165s0030.1.p, Alyli.0064s0064.1.p, Alyli.0060s0091.1.p, Alyli.0480s0011.1.p, Alyli.0480s0013.1.p, Alyli.0060s0093.1.p, Alyli.0060s0094.1.p, Alyli.0314s0007.1.p
#> 4                                                                                                                                                                                                                                  <NA>
#> 5                                                                                                          Alyli.0124s0012.1.p, Alyli.0256s0040.1.p, Alyli.0354s0029.1.p, Alyli.0029s0188.1.p, Alyli.0354s0026.1.p, Alyli.0029s0198.1.p
#> 6                                                                                                                                                                                                                   Alyli.0340s0022.1.p
#>                                       Alyrata_384_v2.1.protein_primaryTranscriptOnly
#> 1                                                                       AL2G12510.t1
#> 2                                                                               <NA>
#> 3 AL3G53950.t1, AL4G11220.t1, AL3G39250.t1, AL4G11250.t1, AL4G11240.t1, AL7G10750.t1
#> 4                                                                       AL4G23510.t1
#> 5                                                         AL1G66210.t1, AL1G66240.t1
#> 6                                                                       AL5G46240.t1
#>   Athaliana_447_Araport11.protein_primaryTranscriptOnly
#> 1                                                  <NA>
#> 2                                                  <NA>
#> 3                 AT5G44220.1, AT2G21930.2, AT4G38870.1
#> 4                              AT2G28960.2, AT2G29000.1
#> 5                                                  <NA>
#> 6                                           AT3G63380.1
#>   Boleraceacapitata_446_v1.0.protein_primaryTranscriptOnly
#> 1                                                     <NA>
#> 2                                                     <NA>
#> 3                                     Bol044742, Bol034450
#> 4                                                     <NA>
#> 5                                                Bol016293
#> 6                                                Bol014472
#>                                                             BrapaFPsc_277_v1.3.protein_primaryTranscriptOnly
#> 1                                                                                                       <NA>
#> 2                                                                                                       <NA>
#> 3 Brara.D01294.1.p, Brara.D01332.1.p, Brara.K01615.1.p, Brara.H01839.1.p, Brara.C00771.1.p, Brara.C00772.1.p
#> 4                                                                                                       <NA>
#> 5                                                       Brara.A02274.1.p, Brara.I01636.1.p, Brara.I01639.1.p
#> 6                                                                                                       <NA>
#>   Bstricta_278_v1.2.protein_primaryTranscriptOnly
#> 1        Bostr.3877s0050.1.p, Bostr.9600s0036.1.p
#> 2                            Bostr.10040s0168.1.p
#> 3       Bostr.5022s0041.1.p, Bostr.25542s0067.1.p
#> 4      Bostr.27991s0024.1.p, Bostr.26833s0919.1.p
#> 5       Bostr.30057s0044.1.p, Bostr.8169s0032.1.p
#> 6                            Bostr.13158s0340.1.p
#>                                                     Camplexicaulis_470_v1.1.protein_primaryTranscriptOnly
#> 1                                                                                     Caamp.0303s1189.1.p
#> 2                                           Caamp.0054s0002.1.p, Caamp.0188s0297.1.p, Caamp.0015s0113.1.p
#> 3 Caamp.1039s0084.1.p, Caamp.1039s0083.1.p, Caamp.0059s0037.1.p, Caamp.1037s0277.1.p, Caamp.1036s0105.1.p
#> 4                                                                Caamp.1033s0379.1.p, Caamp.0051s0002.1.p
#> 5 Caamp.0007s0452.1.p, Caamp.0011s0077.1.p, Caamp.0007s0478.1.p, Caamp.0007s0430.1.p, Caamp.0011s0020.1.p
#> 6                                                                Caamp.0023s0623.1.p, Caamp.0065s0233.1.p
#>                                                                                                                                                                 Cgrandiflora_266_v1.1.protein_primaryTranscriptOnly
#> 1                                                                                                                                                                                                              <NA>
#> 2                                                                                                                                                                                                              <NA>
#> 3 Cagra.7127s0001.1.p, Cagra.27606s0001.1.p, Cagra.4441s0003.1.p, Cagra.0964s0003.1.p, Cagra.0026s0019.1.p, Cagra.4441s0001.1.p, Cagra.9686s0001.1.p, Cagra.0964s0001.1.p, Cagra.1850s0028.1.p, Cagra.1383s0067.1.p
#> 4                                                                                                                                                                                               Cagra.1960s0012.1.p
#> 5                                                                                                                                                    Cagra.10303s0003.1.p, Cagra.4118s0002.1.p, Cagra.4118s0001.1.p
#> 6                                                                                                                                                                                               Cagra.0664s0067.1.p
#>                                    Chispanica_488_v1.1.protein_primaryTranscriptOnly
#> 1                                                                               <NA>
#> 2                                                                Crahi.0165s0003.1.p
#> 3 Crahi.0782s0016.1.p, Crahi.0782s0015.1.p, Crahi.0555s0018.1.p, Crahi.0695s0009.1.p
#> 4                                                                               <NA>
#> 5                                           Crahi.0085s0014.1.p, Crahi.0131s0013.1.p
#> 6                                                                Crahi.1448s0011.1.p
#>                                     Cmaritima_481_v1.1.protein_primaryTranscriptOnly
#> 1                                                                Camar.6671s0001.1.p
#> 2                                                                               <NA>
#> 3 Camar.0266s0009.1.p, Camar.0224s0042.1.p, Camar.8779s0001.1.p, Camar.8779s0002.1.p
#> 4                                                                Camar.0720s0001.1.p
#> 5                      Camar.6467s0001.1.p, Camar.3298s0001.1.p, Camar.2317s0003.1.p
#> 6                                                                Camar.1118s0007.1.p
#>                                                                                                                                                                                                                                                        Crubella_474_v1.1.protein_primaryTranscriptOnly
#> 1                                                                                                                                                                                                                                                                                                 <NA>
#> 2                                                                                                                                                                                                                                                                                                 <NA>
#> 3 Carub.0004s0096.1.p, Carub.0004s0099.1.p, Carub.0004s0102.1.p, Carub.0004s0105.1.p, Carub.0004s0106.1.p, Carub.0004s0538.1.p, Carub.0004s0536.1.p, Carub.0004s0535.1.p, Carub.0004s0537.1.p, Carub.0003s2443.1.p, Carub.0004s0100.1.p, Carub.0004s0107.1.p, Carub.0004s0104.1.p, Carub.0007s2700.1.p
#> 4                                                                                                                                                                                                                                        Carub.0004s0997.1.p, Carub.0004s0995.1.p, Carub.0004s0999.1.p
#> 5                                                                                                                                                                                                                                        Carub.0002s0677.1.p, Carub.0002s0766.1.p, Carub.0002s0765.1.p
#> 6                                                                                                                                                                                                                                                                                                 <NA>
#>                                                          Cviolacea_585_v2.1.protein_primaryTranscriptOnly
#> 1                                                                                                    <NA>
#> 2                                                                                     Clevi.0002s0457.1.p
#> 3                                                                Clevi.0002s0577.1.p, Clevi.0002s0578.1.p
#> 4 Clevi.0006s2387.1.p, Clevi.0006s2389.1.p, Clevi.0006s2391.1.p, Clevi.0006s2392.1.p, Clevi.0006s2390.1.p
#> 5                                                                Clevi.0080s0028.1.p, Clevi.0023s0216.1.p
#> 6                                                                                     Clevi.0009s0601.1.p
#>                                                                                                  Dsophioides_482_v1.1.protein_primaryTranscriptOnly
#> 1                                                                                                                               Desop.0215s0101.1.p
#> 2                                                                                                                                              <NA>
#> 3 Desop.0054s0135.1.p, Desop.0215s0284.1.p, Desop.0041s0048.1.p, Desop.0041s0045.1.p, Desop.0215s0285.1.p, Desop.0215s0283.1.p, Desop.0215s0282.1.p
#> 4                                                                                                                                              <NA>
#> 5                                                                                     Desop.0220s0005.1.p, Desop.0248s0096.1.p, Desop.0248s0089.1.p
#> 6                                                                                                                               Desop.0211s1279.1.p
#>                                                                                                                 Dstrictus_582_v2.1.protein_primaryTranscriptOnly
#> 1                                                                                                                                                           <NA>
#> 2                                                                                                                                          Distr.0009s129000.1.p
#> 3                                                                         Distr.0012s19927.1.p, Distr.0012s17127.1.p, Distr.0022s34400.1.p, Distr.0026s12700.1.p
#> 4 Distr.0010s115200.1.p, Distr.0010s114700.1.p, Distr.0010s114100.1.p, Distr.0010s114900.1.p, Distr.0010s115400.1.p, Distr.0010s77700.1.p, Distr.0010s114200.1.p
#> 5                                                                                             Distr.0009s114600.1.p, Distr.0008s51300.1.p, Distr.0009s110300.1.p
#> 6                                                                                                                                            Distr.0006s2300.1.p
#>                  Esalsugineum_173_v1.0.protein_primaryTranscriptOnly
#> 1                                                               <NA>
#> 2                                                    Thhalv10008885m
#> 3 Thhalv10000202m, Thhalv10000190m, Thhalv10000505m, Thhalv10009644m
#> 4                                   Thhalv10017541m, Thhalv10016227m
#> 5                                   Thhalv10023834m, Thhalv10029493m
#> 6                                                    Thhalv10005762m
#>                                                                               Esyriacum_483_v1.1.protein_primaryTranscriptOnly
#> 1                                                                                                          Eusyr.0117s0372.1.p
#> 2                                                                                                          Eusyr.0018s0571.1.p
#> 3 Eusyr.0134s0602.1.p, Eusyr.0022s0134.1.p, Eusyr.0134s0603.1.p, Eusyr.0022s0133.1.p, Eusyr.0036s0079.1.p, Eusyr.0133s0074.1.p
#> 4                                           Eusyr.0008s0364.1.p, Eusyr.0020s0194.1.p, Eusyr.0008s0362.1.p, Eusyr.0019s0001.1.p
#> 5                                                                Eusyr.0048s0006.1.p, Eusyr.0119s0034.1.p, Eusyr.0119s0033.1.p
#> 6                                                                                                          Eusyr.0016s0019.1.p
#>                                                                                                                                                                                                                                                                                                Evesicaria_487_v1.1.protein_primaryTranscriptOnly
#> 1                                                                                                                                                                                                                                                                                                                                           <NA>
#> 2                                                                                                                                                                                                                                                                                                       Eruve.4404s0002.1.p, Eruve.0725s0026.1.p
#> 3 Eruve.0033s0029.1.p, Eruve.4723s0002.1.p, Eruve.1058s0012.1.p, Eruve.3424s0003.1.p, Eruve.2259s0007.1.p, Eruve.0033s0030.1.p, Eruve.4723s0003.1.p, Eruve.6387s0001.1.p, Eruve.0053s0038.1.p, Eruve.0053s0037.1.p, Eruve.0050s0006.1.p, Eruve.0353s0014.1.p, Eruve.0050s0005.1.p, Eruve.0353s0013.1.p, Eruve.0050s0007.1.p, Eruve.0353s0015.1.p
#> 4                                                                                                                                                                                                                                                                                                       Eruve.1151s0002.1.p, Eruve.0271s0002.1.p
#> 5                                                                                                                                                                                                                                                             Eruve.8324s0002.1.p, Eruve.0513s0008.1.p, Eruve.2726s0014.1.p, Eruve.5754s0001.1.p
#> 6                                                                                                                                                                                                                                                                                                                            Eruve.0058s0013.1.p
#>                   Iamara_485_v1.1.protein_primaryTranscriptOnly
#> 1                                                          <NA>
#> 2                      Ibeam.0043s0014.1.p, Ibeam.6308s0002.1.p
#> 3                                           Ibeam.0795s0031.1.p
#> 4 Ibeam.0194s0002.1.p, Ibeam.5235s0001.1.p, Ibeam.0300s0005.1.p
#> 5                      Ibeam.4441s0001.1.p, Ibeam.0061s0043.1.p
#> 6                                           Ibeam.1101s0005.1.p
#>                                                                                                                                                                                         Itinctoria_475_v1.1.protein_primaryTranscriptOnly
#> 1                                                                                                                                                                                                                     Isati.2167s0003.1.p
#> 2                                                                                       Isati.0196s0023.1.p, Isati.0173s0037.1.p, Isati.0173s0039.1.p, Isati.4321s0004.1.p, Isati.1180s0013.1.p, Isati.2208s0017.1.p, Isati.0515s0018.1.p
#> 3                        Isati.2844s0003.1.p, Isati.0755s0009.1.p, Isati.4212s0010.1.p, Isati.4212s0009.1.p, Isati.0590s0005.1.p, Isati.0996s0008.1.p, Isati.1578s0015.1.p, Isati.6137s0002.1.p, Isati.0033s0030.1.p, Isati.1971s0016.1.p
#> 4  Isati.0016s0022.1.p, Isati.10929s0001.1.p, Isati.0447s0006.1.p, Isati.1678s0009.1.p, Isati.6700s0004.1.p, Isati.0739s0001.1.p, Isati.0929s0015.1.p, Isati.2436s0018.1.p, Isati.4307s0002.1.p, Isati.3568s0001.1.p, Isati.4307s0004.1.p
#> 5 Isati.8225s0001.1.p, Isati.2999s0002.1.p, Isati.2265s0007.1.p, Isati.4790s0003.1.p, Isati.0045s0034.1.p, Isati.14169s0001.1.p, Isati.4314s0004.1.p, Isati.2925s0007.1.p, Isati.18580s0001.1.p, Isati.1835s0001.1.p, Isati.7241s0003.1.p
#> 6                                                                                                                                                                                                                     Isati.1303s0013.1.p
#>   Lannua_476_v1.1.protein_primaryTranscriptOnly
#> 1                                          <NA>
#> 2      Luann.0029s0247.1.p, Luann.0053s0157.1.p
#> 3                                          <NA>
#> 4                                          <NA>
#> 5                                          <NA>
#> 6                           Luann.0568s0015.1.p
#>   Lsativum_478_v1.1.protein_primaryTranscriptOnly
#> 1                                            <NA>
#> 2                                            <NA>
#> 3                                            <NA>
#> 4                                            <NA>
#> 5        Lesat.0082s0065.1.p, Lesat.0017s0067.1.p
#> 6        Lesat.0063s0077.1.p, Lesat.0025s0247.1.p
#>                                                          Mmaritima_477_v1.1.protein_primaryTranscriptOnly
#> 1                                                                                     Mamar.0093s0032.1.p
#> 2                                                                                                    <NA>
#> 3 Mamar.0021s0084.1.p, Mamar.0068s0003.1.p, Mamar.0068s0001.1.p, Mamar.0068s0002.1.p, Mamar.0068s0004.1.p
#> 4                                                                                     Mamar.0087s0377.1.p
#> 5                                           Mamar.0048s0140.1.p, Mamar.0105s0048.1.p, Mamar.0105s0047.1.p
#> 6                                                                                     Mamar.0015s0294.1.p
#>             Mperfoliatum_583_v2.1.protein_primaryTranscriptOnly
#> 1                                           Myper.0014s0450.1.p
#> 2                                           Myper.0014s0472.1.p
#> 3                      Myper.0012s0510.1.p, Myper.0012s0415.1.p
#> 4 Myper.0003s0342.1.p, Myper.0003s0345.1.p, Myper.0003s0346.1.p
#> 5 Myper.0014s0283.1.p, Myper.0014s0241.1.p, Myper.0014s0312.1.p
#> 6                                           Myper.0082s0039.1.p
#>                                                                                                                                             Rislandica_473_v1.1.protein_primaryTranscriptOnly
#> 1                                                                                                                                                                         Roisl.0010s0123.1.p
#> 2                                                                                                                                                                         Roisl.0214s0596.1.p
#> 3 Roisl.0016s0356.1.p, Roisl.0046s0648.1.p, Roisl.0046s0647.1.p, Roisl.0209s0557.1.p, Roisl.0209s0558.1.p, Roisl.0046s0687.1.p, Roisl.0075s0210.1.p, Roisl.0075s0209.1.p, Roisl.0209s0072.1.p
#> 4                                                                                                          Roisl.0082s0688.1.p, Roisl.0082s0678.1.p, Roisl.0082s0684.1.p, Roisl.0082s0683.1.p
#> 5                                                                                                                               Roisl.0008s0033.1.p, Roisl.0008s0034.1.p, Roisl.0008s0037.1.p
#> 6                                                                                                                                                                         Roisl.0018s0479.1.p
#>                    Salba_584_v3.1.protein_primaryTranscriptOnly
#> 1                                           Sialb.0003s0107.1.p
#> 2                                           Sialb.0897s0019.1.p
#> 3                      Sialb.0004s0218.1.p, Sialb.1306s0020.1.p
#> 4                      Sialb.0263s0051.1.p, Sialb.0263s0049.1.p
#> 5 Sialb.0102s0081.1.p, Sialb.0083s0107.1.p, Sialb.1407s0016.1.p
#> 6                      Sialb.2258s0007.1.p, Sialb.1674s0011.1.p
#>   Sparvula_574_v2.2.protein_primaryTranscriptOnly
#> 1                                            <NA>
#> 2                        Sp1g09310.1, Sp2g04250.1
#> 3                                     Sp6g06960.1
#> 4                                     Sp4g11810.1
#> 5           Sp2g05750.1, Sp2g06290.1, Sp2g05530.1
#> 6                                     Sp5g00170.1
#>                                                                                Spinnata_486_v1.1.protein_primaryTranscriptOnly
#> 1                                                                                                          Stapi.3691s0002.1.p
#> 2                                                                                                          Stapi.5707s0002.1.p
#> 3 Stapi.0026s0005.1.p, Stapi.2138s0001.1.p, Stapi.2092s0004.1.p, Stapi.4624s0004.1.p, Stapi.2863s0001.1.p, Stapi.7828s0001.1.p
#> 4                                                                                     Stapi.7251s0001.1.p, Stapi.1778s0002.1.p
#> 5                                           Stapi.0026s0004.1.p, Stapi.1284s0007.1.p, Stapi.2700s0001.1.p, Stapi.5887s0001.1.p
#> 6                                                                                     Stapi.0594s0011.1.p, Stapi.3122s0003.1.p
#>                                                           Tarvense_479_v1.1.protein_primaryTranscriptOnly
#> 1                                                                                                    <NA>
#> 2                                                                Thlar.0384s0011.1.p, Thlar.0006s0757.1.p
#> 3 Thlar.0024s0208.1.p, Thlar.0003s0240.1.p, Thlar.0028s0069.1.p, Thlar.0008s0489.1.p, Thlar.0008s0488.1.p
#> 4                                           Thlar.0011s0400.1.p, Thlar.0011s0392.1.p, Thlar.0011s0394.1.p
#> 5                                                                Thlar.0212s0013.1.p, Thlar.0335s0003.1.p
#> 6                                                                                     Thlar.0021s0018.1.p
#>   N_0 N_1 pvalueFisherSelf-incompatible oddsRatioFisherSelf-incompatible
#> 1  11   0                    0.02374582                              Inf
#> 2  15   1                    0.01842505                        13.239525
#> 3  20   4                    0.04615385                              Inf
#> 4  17   2                    0.02775920                         9.918203
#> 5  20   4                    0.04615385                              Inf
#> 6  20   4                    0.04615385                              Inf
#>          Gene Nannotations             X2
#> 1        <NA>           NA           <NA>
#> 2        <NA>           NA           <NA>
#> 3 AT5G44220.1            3 protein_coding
#> 4 AT2G29000.1            3 protein_coding
#> 5        <NA>           NA           <NA>
#> 6 AT3G63380.1            3 protein_coding
#>                                                                                       X3
#> 1                                                                                   <NA>
#> 2                                                                                   <NA>
#> 3                                                                   F-box family protein
#> 4                                      Leucine-rich repeat protein kinase family protein
#> 5                                                                                   <NA>
#> 6 ATPase E1-E2 type family protein / haloacid dehalogenase-like hydrolase family protein
#>     X4
#> 1 <NA>
#> 2 <NA>
#> 3 <NA>
#> 4 <NA>
#> 5 <NA>
#> 6 <NA>
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            X5
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        <NA>
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        <NA>
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    F-box family protein; CONTAINS InterPro DOMAIN/s: F-box domain, cyclin-like (InterPro:IPR001810), F-box domain, Skp2-like (InterPro:IPR022364); BEST Arabidopsis thaliana protein match is: F-box and associated interaction domains-containing protein (TAIR:AT4G38870.1); Has 1045 Blast hits to 1017 proteins in 38 species: Archae - 0; Bacteria - 0; Metazoa - 0; Fungi - 0; Plants - 1045; Viruses - 0; Other Eukaryotes - 0 (source: NCBI BLink).
#> 4                                                                                                                                                                                                                                                                                                                                                                              Leucine-rich repeat protein kinase family protein; FUNCTIONS IN: protein serine/threonine kinase activity, protein kinase activity, ATP binding; INVOLVED IN: protein amino acid phosphorylation; LOCATED IN: endomembrane system; CONTAINS InterPro DOMAIN/s: Protein kinase, ATP binding site (InterPro:IPR017441), Protein kinase, catalytic domain (InterPro:IPR000719), Serine/threonine-protein kinase domain (InterPro:IPR002290), Tyrosine-protein kinase, catalytic domain (InterPro:IPR020635), Serine/threonine-protein kinase-like domain (InterPro:IPR017442), Serine/threonine-protein kinase, active site (InterPro:IPR008271), Protein kinase-like domain (InterPro:IPR011009); BEST Arabidopsis thaliana protein match is: Leucine-rich repeat protein kinase family protein (TAIR:AT2G28960.1); Has 35333 Blast hits to 34131 proteins in 2444 species: Archae - 798; Bacteria - 22429; Metazoa - 974; Fungi - 991; Plants - 531; Viruses - 0; Other Eukaryotes - 9610 (source: NCBI BLink).
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        <NA>
#> 6 ATPase E1-E2 type family protein / haloacid dehalogenase-like hydrolase family protein; FUNCTIONS IN: calcium-transporting ATPase activity, calmodulin binding; INVOLVED IN: calcium ion transport, cation transport, metabolic process, ATP biosynthetic process; LOCATED IN: membrane; EXPRESSED IN: 11 plant structures; EXPRESSED DURING: LP.06 six leaves visible, LP.04 four leaves visible, 4 anthesis, petal differentiation and expansion stage; CONTAINS InterPro DOMAIN/s: ATPase, P-type, ATPase-associated domain (InterPro:IPR008250), ATPase, P-type, calcium-transporting, PMCA-type (InterPro:IPR006408), ATPase, P-type, H+ transporting proton pump (InterPro:IPR000695), ATPase, P-type cation-transporter, N-terminal (InterPro:IPR004014), Haloacid dehalogenase-like hydrolase (InterPro:IPR005834), ATPase, P-type, K/Mg/Cd/Cu/Zn/Na/Ca/Na/H-transporter (InterPro:IPR001757), ATPase, P-type phosphorylation site (InterPro:IPR018303), ATPase, P-type cation-transporter, C-terminal (InterPro:IPR006068); BEST Arabidopsis thaliana protein match is: ATPase E1-E2 type family protein / haloacid dehalogenase-like hydrolase family protein (TAIR:AT3G22910.1); Has 46574 Blast hits to 34715 proteins in 3214 species: Archae - 909; Bacteria - 32032; Metazoa - 4086; Fungi - 2666; Plants - 2162; Viruses - 3; Other Eukaryotes - 4716 (source: NCBI BLink).

# OG of the gene AT5G44220.1
GeneID <- c("AT5G44220.1")
OrthoFinderID <- c("Athaliana_447_Araport11.protein_primaryTranscriptOnly")
AT5G44220_OG <- obtain_OG_from_gene(GeneDiscoveRobject, GeneID, OrthoFinderID)
#> Error in filter(., str_detect(pattern = GeneID[1], across(OrthoFinderID[1]))): ℹ In argument: `str_detect(pattern = GeneID[1],
#>   across(OrthoFinderID[1]))`.
#> ℹ In row 1.
#> Caused by error in `across()`:
#> ℹ In argument: `OrthoFinderID[1]`.
#> Caused by error:
#> ! object 'OrthoFinderID' not found
AT5G44220_OG # Identify the Orthologous Group (OG) of the gene
#> Error in eval(expr, envir, enclos): object 'AT5G44220_OG' not found